requirement.txt
albumentations==0.4.3
opencv-python==4.1.2.30
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
torchmetrics==0.6
kornia==0.5.10
invisible-watermark
diffusers
transformers==4.19.2
nltk
ftfy
six